Cleaned up and finalized some classes. Prepared to integreate CMConnection into the other classes.
