+++ src/main/java/inc/vareli/crusman/UI/CrusMan.java
@@ -1,11 +1,7 @@
 package inc.vareli.crusman.UI;
 
-import inc.vareli.crusman.data.*;
-
 public class CrusMan {
 	public static void main(String[] args) {
-			Ship ship;
-			Trip trip;
 			//this is the class that launches the GUI for the user to
 			//interface with the system.
 			//in the end this will be in a different package from the other
