+++ src/main/java/inc/vareli/crusman/databases/CostType.java
@@ -1,0 +1,5 @@
+package inc.vareli.crusman.databases;
+
+public interface CostType {
+
+}
